Struct automaton::dfa::core::DFABuilder

#[must_use = "builders need to be built with .finalize() method"]

pub struct DFABuilder { /* fields omitted */ }

The DFABuilder follows the builder pattern and allows to create a Deterministic Finite Automaton. The builder is moved at each call so it is necessary to bind to a new variable the return value for each function of the builder.

Errors

Return an error if the starting state is not specified.

Return an error if the final states are not specified.

Examples

extern crate automaton;

use automaton::dfa::core::*;
use std::error::Error;
 
fn main() {
    // (abc)*
    let dfa = DFABuilder::new()
        .add_start(0)
        .add_final(0)
        .add_transition('a', 0, 1)
        .add_transition('b', 1, 2)
        .add_transition('c', 2, 0)
        .finalize();
}

extern crate automaton;

use automaton::dfa::core::*;
use std::error::Error;
 
fn main() {
    let dfa = DFABuilder::new()
        .add_start(4)
        .add_transition('t', 0, 1)
        .finalize();
    match dfa {
        Err(DFAError::MissingFinalStates) => assert!(true),
        _ => assert!(false),
    }
}

extern crate automaton;

use automaton::dfa::core::*;
use std::error::Error;
 
fn main() {
    let dfa = DFABuilder::new()
        .add_start(4)
        .add_transition('t', 0, 1)
        .add_transition('t', 0, 2)
        .finalize();
    match dfa {
        Err(DFAError::DuplicatedTransition(symb,src)) => assert!((symb,src)==('t',0)),
        _ => assert!(false),
    }
}

extern crate automaton;

use automaton::dfa::core::*;
use std::error::Error;
 
fn main() {
    let dfa = DFABuilder::new()
        .add_final(4)
        .add_transition('t', 0, 1)
        .finalize();
    match dfa {
        Err(DFAError::MissingStartingState) => assert!(true),
        _ => assert!(false),
    }
}

Methods

impl DFABuilder

fn new() -> Result<DFABuilder>

Creates a new DFABuilder.