Trait automaton::dfa::core::DFABuilding

pub trait DFABuilding {
    type Type: Sized;
    type Builder: Sized;
    fn add_start(self, state: usize) -> Result<Self::Builder>;

    fn add_final(self, state: usize) -> Result<Self::Builder>;

    fn add_transition(
        self, 
        symb: char, 
        src: usize, 
        dest: usize
    ) -> Result<Self::Builder>;

    fn finalize(self) -> Result<Self::Type>;
}

DFABuilding is the trait assiociated to the DFABuilder type. Each DFABuilder should implement DFABuilding trait.

DFABuilder can generate some errors during the building stage. For instance, one could try to insert a transition with two different destination states.

Errors

If self contains a DFAerror then each function should transfer this error.

Associated Types

type Type: Sized

Type of DFA returned

type Builder: Sized

Type of the builder used

Required Methods

fn add_start(self, state: usize) -> Result<Self::Builder>

Add a starting state to the DFA.

Errors

In the futur will return a DFAError::DuplicatedStartingState if two starting states are added.

fn add_final(self, state: usize) -> Result<Self::Builder>

Add a final state to the DFA.

fn add_transition(
self,
symb: char,
src: usize,
dest: usize
) -> Result<Self::Builder>

Add a transition to the DFA.

Errors

Return a DFAError::DuplicatedTransition(symb,src) if a transtion with the same symb and src has already been inserted, even if the destination state is the same.

fn finalize(self) -> Result<Self::Type>

Finalize the building of the DFA.

Errors

Return a DFAError::MissingStartingState if no starting state is specified.

Return a DFAError::MissingFinalStates if no final state is specified.