Module automaton::e_nfa::core

e_nfa core api

Structs

ENFA
The type ENFA represents a NonDeterministic Finite Automaton. The transitions of the automatonn are stored in a hashtable.

ENFABuilder
The ENFABuilder follows the builder pattern and allows to create a Deterministic Finite Automaton. The builder is moved at each call so it is necessary to bind to a new variable the return value for each function of the builder.

Enums

ENFAError
The ENFAError type.

Traits

ENFABuilding
ENFABuilding is the trait assiociated to the ENFABuilder type. Each ENFABuilder should implement ENFABuilding trait.

Type Definitions

Result
Alias for result::Result.