Trait automaton::e_nfa::core::ENFABuilding

pub trait ENFABuilding {
    fn add_start(self, state: usize) -> Result<ENFABuilder>;

    fn add_final(self, state: usize) -> Result<ENFABuilder>;

    fn add_transition(
        self, 
        symb: char, 
        src: usize, 
        dest: usize
    ) -> Result<ENFABuilder>;

    fn add_e_transition(self, src: usize, dest: usize) -> Result<ENFABuilder>;

    fn finalize(self) -> Result<ENFA>;
}

ENFABuilding is the trait assiociated to the ENFABuilder type. Each ENFABuilder should implement ENFABuilding trait.

ENFABuilder can generate some errors during the building stage. For instance, one could try to insert a transition with two different destination states.

Errors

If self contains a ENFAerror then each function should transfer this error.

Required Methods

fn add_start(self, state: usize) -> Result<ENFABuilder>

Add a starting state to the ENFA.

Errors

In the futur will return a ENFAError::DuplicatedStartingState if two starting states are added.

fn add_final(self, state: usize) -> Result<ENFABuilder>

Add a final state to the ENFA.

fn add_transition(
self,
symb: char,
src: usize,
dest: usize
) -> Result<ENFABuilder>

Add a transition to the ENFA.

fn add_e_transition(self, src: usize, dest: usize) -> Result<ENFABuilder>

Add an epsilon transition to the ENFA.

fn finalize(self) -> Result<ENFA>

Finalize the building of the ENFA.

Errors

Return a ENFAError::MissingStartingState if no starting state is specified.

Return a ENFAError::MissingFinalStates if no final state is specified.