Struct automaton::nfa::core::NFA

pub struct NFA { /* fields omitted */ }

The type NFA represents a NonDeterministic Finite Automaton. The transitions of the automatonn are stored in a hashtable.

Methods

impl NFA

fn test(&self, input: &str) -> bool

Test if an input string is a word of the language defined by the NFA.

Examples

extern crate automaton;

use automaton::nfa::core::*;
use std::error::Error;
 
fn main() {
    // (abc)*
    let nfa = NFABuilder::new()
        .add_start(0)
        .add_final(3)
        .add_final(0)
        .add_transition('a', 0, 1)
        .add_transition('b', 1, 2)
        .add_transition('c', 2, 3)
        .add_transition('a', 3, 1)
        .finalize();
    match nfa {
        Ok(nfa) => {
           assert!(nfa.test("abc"));
           assert!(nfa.test(""));
           assert!(!nfa.test("a"));
           assert!(!nfa.test("ab"));
           assert!(!nfa.test("abca"));
           assert!(!nfa.test("abcab"));
           assert!(nfa.test("abcabcabc"));
        },
        Err(e) => println!("{}", e),
    }
}