Struct automaton::nfa::core::NFABuilder

pub struct NFABuilder { /* fields omitted */ }

The NFABuilder follows the builder pattern and allows to create a Deterministic Finite Automaton. The builder is moved at each call so it is necessary to bind to a new variable the return value for each function of the builder.

Errors

Return an error if the starting state is not specified.

Return an error if the final states are not specified.

Examples

extern crate automaton;

use automaton::nfa::core::*;
use std::error::Error;
 
fn main() {
    // (abc)*
    let nfa = NFABuilder::new()
        .add_start(0)
        .add_final(3)
        .add_final(0)
        .add_transition('a', 0, 1)
        .add_transition('b', 1, 2)
        .add_transition('c', 2, 3)
        .add_transition('a', 3, 1)
        .finalize();
}

extern crate automaton;

use automaton::nfa::core::*;
use std::error::Error;
 
fn main() {
    let nfa = NFABuilder::new()
        .add_start(4)
        .add_transition('t', 0, 1)
        .finalize();
    match nfa {
        Err(NFAError::MissingFinalStates) => assert!(true),
        _ => assert!(false),
    }
}

extern crate automaton;

use automaton::nfa::core::*;
use std::error::Error;
 
fn main() {
    let nfa = NFABuilder::new()
        .add_final(4)
        .add_transition('t', 0, 1)
        .finalize();
    match nfa {
        Err(NFAError::MissingStartingState) => assert!(true),
        _ => assert!(false),
    }
}

Methods

impl NFABuilder

fn new() -> Result<NFABuilder>

Creates a new NFABuilder.