Enum automaton::nfa::reader::NFAReaderError

pub enum NFAReaderError {
    MissingStartingState,
    MissingFinalStates,
    IncompleteTransition(usize),
    IllformedTransition(usize),
    NFA(NFAError, usize),
    Io(Error),
    Parse(ParseIntError, usize),
}

Type NFAReaderError describes the list of errors that can occur during the parsing of a NFA file.

Variants
MissingStartingState

Error MissingStartingState means the file does not contains the starting state.

MissingFinalStates

Error MissingFinalStates means the file does not contains the list of final states.

IncompleteTransition(usize)

Error IncompleteTransition means the transition on the specified line does not contain one of these elements: symbol, source state, destination state.

IllformedTransition(usize)

Error IllformedTransition means the transition contains to much elements or that the symbole is composed with modre than two characters.

NFA(NFAError, usize)

Error NFA encapsules the error specific to the NFA building process (no final states,...).

Io(Error)

Error Io is relative to the input errors (the file does not exist, the file can not be read,...à.

Parse(ParseIntError, usize)

Error Parse is relative to the parsing errors (a state is an intger).