Struct automaton::nfa::reader::NFAReader

pub struct NFAReader;

Struct NFAReader is an empty structure that builds a NFA from a file or from a &str.

Methods

impl NFAReader

fn new_from_file<P: AsRef<Path>>(file_path: P) -> Result<NFA>

Reads a NFA from a file.

Description

file_path - The path to the file that contains the NFA.

Examples

extern crate automaton;

use automaton::nfa::reader::*;
use std::error::Error;
 
fn main() {
    let nfa = NFAReader::new_from_file("nfa.txt");
    match nfa {
        Ok(nfa) => {
           // Do stuff with the nfa
        },
        Err(e) => println!("{}", e),
    }
}

fn new_from_string(nfa: &str) -> Result<NFA>

Reads a NFA from a &str.

Description

nfa - The string representation of the NFA.

Examples

extern crate automaton;

use automaton::nfa::reader::*;
use std::error::Error;
 
fn main() {
    // (abc)*
    let nfa =
        "0 1\n\
         0 3\n\
         a 0 1\n\
         b 1 2\n\
         c 2 3\n\
         a 3 0";
    let nfa = NFAReader::new_from_string(nfa);
    match nfa {
        Ok(nfa) => {
           // Do stuff with the nfa
        },
        Err(e) => println!("{}", e),
    }
}